Rename namespace utilities.
