+++ packages/field-scripts/bin/develop-module/mock-server.js
@@ -1,5 +1,5 @@
 import http from 'http';
-import { FM_API_ENDPOINT, resolveModulePathname } from '../../lib/constants.js';
+import { FM_API_ENDPOINT, resolveModulePathname } from '../../namespaces/constants.js';
 import { fmtScript, fmtJsonApi } from '../format.js';
 
 const trimPath = p => (typeof p === 'string' ? p : '').replaceAll(/^\/|\/$/g, '');
