+++ packages/field-scripts/namespaces/index.js
